+++ djangoapp/migrations/0002_portfoliosnapshot.py
@@ -15,8 +15,8 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('created_at', models.DateTimeField(auto_now_add=True)),
-                ('total', models.FloatField()),
-                ('deposit', models.FloatField()),
+                ('total', models.IntegerField()),
+                ('deposit', models.IntegerField()),
             ],
         ),
     ]
